quick-settings: try harder to retrieve the list of profiles
If quick-settings is displayed too early on startup,
phoneui_utils_sound_profile_list() fails, resulting in an error message
and an empty profile combo that stays empty until phoneuid is restarted.

This patch changes the code to keep retrying after a given timeout
(currently 2 sec) until the profile list is received.

As an added bonus, the current profile is queried and selected into the
list combo once the profile list is filled, so it should always reflect
the current profile. The combo box is disabled until the profile list is
filled.

Signed-off-by: David Kozub <zub@linux.fjfi.cvut.cz>