message-list-view: use the list scroller events for retriving messages
This patch uses the new "scroll,edge,bottom" and "scroll,edge,top" genlist
callbacks for populating the message view.
By default it loads max MSG_PAGE_SIZE messages per page, then when the message
list is scrolled to bottom it tryies to load more messages in the list; when the
bottom top edge is reached by scrolling up, the list tries to fetch the previous
messages (if they exist).

Visual feedback (progress) to be done.

Signed-off-by: Klaus Kurzmann <mok@fluxnetz.de>