message-list-view: workaround to make the list reload the contact name
This happens when no "Name" value is found in the message hash_table, due (I guess)
to some race conditions of the genlist which causes not to save the "Name" value
for ever in the hashtable.

This is an ugly workaround (revert if you don't like it) to make the genlist to
lookup for the contact name every time it doesn't find it. Scrolling the list, in
fact, make the name caming back.